backend-ai/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
openai==1.12.0
qdrant-client==1.7.3
langchain==0.1.0
langchain-openai==0.0.2
pypdf==4.0.0
python-docx==1.1.0
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
httpx==0.27.2
